Filter workload forward targets by controller ownership
A label selector alone can match pods the workload does not own (bare
pods with copied labels, overlapping selectors), so forwarding to a
workload could route to a stranger pod. Move the ownership-aware pod
resolution used by log targets into kube/target-pods.ts and use it for
port-forward target resolution too: pods must be controller-owned by
the workload (via its owned ReplicaSets for Deployments).
