+++ server/src/routes/detail.ts
@@ -2,73 +2,14 @@ import { X509Certificate } from 'node:crypto';
 import type { FastifyInstance } from 'fastify';
 import type { KubeObject, LogTargetKind, LogTargetPodsResponse, PodEnvResponse, SecretTlsResponse, TlsCertInfo } from '@kubus/shared';
 import type { AppContext } from '../app.js';
-import type { ClusterHandle } from '../kube/cluster-manager.js';
 import { podContainers } from '../kube/actions.js';
 import { getRolloutHistory } from '../kube/rollout.js';
 import { resolvePodEnv } from '../kube/pod-env.js';
 import { resourcePath } from '../kube/raw-client.js';
+import { resolveTargetPods } from '../kube/target-pods.js';
 import { HttpProblem, sendError } from '../util/errors.js';
 
-interface LabelSelector {
-  matchLabels?: Record<string, string>;
-  matchExpressions?: Array<{ key: string; operator: 'In' | 'NotIn' | 'Exists' | 'DoesNotExist'; values?: string[] }>;
-}
-
 const CERT_BLOCK_RE = /-----BEGIN CERTIFICATE-----[\s\S]*?-----END CERTIFICATE-----/g;
-
-function selectorToString(selector: LabelSelector | undefined): string | undefined {
-  if (!selector) return undefined;
-  const parts = Object.entries(selector.matchLabels ?? {}).map(([k, v]) => `${k}=${v}`);
-  for (const expr of selector.matchExpressions ?? []) {
-    if (expr.operator === 'In') parts.push(`${expr.key} in (${(expr.values ?? []).join(',')})`);
-    else if (expr.operator === 'NotIn') parts.push(`${expr.key} notin (${(expr.values ?? []).join(',')})`);
-    else if (expr.operator === 'Exists') parts.push(expr.key);
-    else if (expr.operator === 'DoesNotExist') parts.push(`!${expr.key}`);
-  }
-  return parts.length ? parts.join(',') : undefined;
-}
-
-async function listPods(handle: ClusterHandle, namespace: string, selector?: string): Promise<KubeObject[]> {
-  const query = new URLSearchParams();
-  if (selector) query.set('labelSelector', selector);
-  const list = await handle.raw.json<{ items?: KubeObject[] }>(resourcePath('', 'v1', 'pods', { namespace, query }));
-  return list.items ?? [];
-}
-
-function owns(obj: KubeObject, uid: string | undefined): boolean {
-  if (!uid) return false;
-  return (obj.metadata.ownerReferences ?? []).some((owner) => owner.uid === uid && owner.controller);
-}
-
-async function resolveLogTargetPods(handle: ClusterHandle, target: KubeObject, kind: LogTargetKind, namespace: string): Promise<KubeObject[]> {
-  if (kind === 'Pod') return [target];
-
-  if (kind === 'Service') {
-    const selector = (target.spec as { selector?: Record<string, string> } | undefined)?.selector;
-    const labelSelector = selectorToString({ matchLabels: selector });
-    return labelSelector ? listPods(handle, namespace, labelSelector) : [];
-  }
-
-  const selector = selectorToString((target.spec as { selector?: LabelSelector } | undefined)?.selector);
-  if (kind === 'Job') {
-    const pods = await listPods(handle, namespace, selector);
-    return pods.filter((pod) => owns(pod, target.metadata.uid));
-  }
-  if (!selector) return [];
-
-  if (kind === 'Deployment') {
-    const query = new URLSearchParams({ labelSelector: selector });
-    const [rsList, pods] = await Promise.all([
-      handle.raw.json<{ items?: KubeObject[] }>(resourcePath('apps', 'v1', 'replicasets', { namespace, query })),
-      listPods(handle, namespace, selector),
-    ]);
-    const rsUids = new Set((rsList.items ?? []).filter((rs) => owns(rs, target.metadata.uid)).map((rs) => rs.metadata.uid));
-    return pods.filter((pod) => (pod.metadata.ownerReferences ?? []).some((owner) => rsUids.has(owner.uid) && owner.controller));
-  }
-
-  const pods = await listPods(handle, namespace, selector);
-  return pods.filter((pod) => owns(pod, target.metadata.uid));
-}
 
 export function registerDetailRoutes(app: FastifyInstance, ctx: AppContext): void {
   app.get<{ Params: { ctx: string }; Querystring: { namespace?: string; name?: string; reveal?: string } }>(
@@ -96,7 +37,7 @@ export function registerDetailRoutes(app: FastifyInstance, ctx: AppContext): voi
       if (!version || !plural || !kind || !namespace || !name) throw new HttpProblem(422, 'group, version, plural, kind, namespace and name are required');
       const handle = ctx.clusters.get(req.params.ctx);
       const target = await handle.raw.json<KubeObject>(resourcePath(group, version, plural, { namespace, name }));
-      const pods = await resolveLogTargetPods(handle, target, kind, namespace);
+      const pods = await resolveTargetPods(handle, target, kind, namespace);
       const response: LogTargetPodsResponse = {
         pods: pods
           .map((pod) => ({
